カダイ

+++ app/Http/Controllers/TasksController.php
@@ -3,20 +3,30 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+
+use Illuminate\Support\Facades\Auth; 
 use App\Models\Task;
 
 class TasksController extends Controller
 {
-    // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::all();
-
-        //　タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);  
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('id', 'asc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            // users.indexビューでそれらを表示
+            return view('users.index', $data);
+        }else{
+            return view('dashboard');
+        }
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -33,6 +43,7 @@ class TasksController extends Controller
     // postでtasks/にアクセスされた場合の「新規登録処理」
     public function store(Request $request)
     {
+        $user = Auth::user();
         // バリデーション
         $request->validate([
             'status' => 'required|max:10',
@@ -43,22 +54,12 @@ class TasksController extends Controller
         $task = new Task;
         $task -> status = $request->status;
         $task -> content = $request->content;
+        $task -> user_id = $user->id;
+        // dd($task->user_id);
         $task -> save();
 
         // トップページへリダイレクトさせる
-        return redirect('/');
-    }
-
-    // getでtasks/（任意のid）にアクセスされた場合の「取得表示処理」
-    public function show($id)
-    {
-        // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-
-        // タスク詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        return redirect('/dashboard');
     }
 
     // getでtasks/（任意のid）/editにアクセスされた場合の「更新画面表示処理」
@@ -90,7 +91,7 @@ class TasksController extends Controller
         $task->save();
 
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('/dashboard');
     }
 
     // deleteでtasks/（任意のid）にアクセスされた場合の「削除処理」
@@ -98,10 +99,17 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/dashboard');
+        }
+        
+        // // タスクを削除
+        // $task->delete();
 
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('/dashboard')
+            ->with('Delete Failed');;
     }
 }
